[clangd] Cache preambles of closed files
When a file is closed, push its preamble to a LRU cache
When a file is opened, try to get the preamble from the LRU cache.

By default store 10 preambles if they are stored on
memory and 1000 if they are stored on disk. That value
can be modified with --keep-preambles=N

Differential Revision: https://reviews.llvm.org/D93873