[clangd] use a compatible preamble for the first AST built
Keep a store of the preambles of all opened filed, plus up to
5 previously used preambles. When building the AST for a TU,
if the preamble for this TU is not yet built, look through
the stored premables to find the best compatible preamble
and use this preamble to speed-up the AST build.

---------

This patch is starting to look like something acceptable, so
I'm sending it for review. At this point I'd like to get some
feedback on this before going further.

Differential Revision: https://reviews.llvm.org/D97417